% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artworks.R
\name{get_artworks}
\alias{get_artworks}
\title{Get artworks}
\usage{
get_artworks(object_ids = NULL, fields = NULL)
}
\arguments{
\item{object_ids}{Artwork object ids based on Artic}

\item{fields}{Fields to be extracted}
}
\value{
A data frame.
}
\description{
Get artworks
}
\examples{
get_artworks(4)
}
